Create data table

Create an interactive data table with download buttons.

create_DT(data, caption = "", scrollY = 400)

Arguments

data

a data object (either a matrix or a data frame)

caption

the table caption; a character vector or a tag object generated from htmltools::tags$caption()

scrollY

Max height (in pixels) before wrapping the table in a vertical scroll bar.

Value

datatables object.

Examples

data("mtcars")
create_DT(mtcars)
#> Loading required namespace: DT