Import precomputed CellTypeDatasets

Import CellTypeDatasets (CTD) generated by EWCE.

"ctd_AIBS"
CTD file derived from adult human cortex scRNA-seq data collected by the Allen Institute for Brain Science (AIBS) Note that this CTD used an early release of the AIBS data that only included samples from human Medial Temporal Gyrus (MTG).
Reference: doi:10.1038/s41586-019-1506-7
Allen Brain Institute website
Source

"ctd_allAIBS"
CTD file derived from adult human cortex scRNA-seq data collected by the Allen Institute for Brain Science (AIBS) Note that this CTD used an later release of the AIBS data that included samples from multiple human brain regions.
Reference: doi:10.1038/s41586-019-1506-7
Allen Brain Institute website
Source

"ctd_allKI"
CTD file with cortex, hippocampus, hypothalamus and midbrain.
Reference: doi:10.1038/s41588-018-0129-5
Source

"ctd_BlueLake2018_FrontalCortexOnly"
CTD file with data from humans.
Reference: doi:10.1038/nbt.4038
Source

"ctd_BlueLake2018_VisualCortexOnly"
CTD file with data from humans.
Reference: doi:10.1038/nbt.4038
Source

"ctd_Saunders"
CTD file with data from humans.
Reference: doi:10.1016/j.cell.2018.07.028
Source

"ctd_DescartesHuman"
CTD file with scRNA-seq data from human embryo across multiple organ systems.
Reference: doi:10.1126/science.aba7721
Source

"ctd_DRONC_human"
CTD file with DRONC-seq data from humans.
Reference: doi:10.1038/nmeth.4407
Source

"ctd_DRONC_mouse"
CTD file with Tasic data from mouse.
Reference: doi:10.1038/nmeth.4407
Reference

"ctd_DivSeq"
CTD file with DivSeq data.
Reference: doi:10.1126/science.aad7038
Source

"ctd_Tasic"
CTD file with Tasic data.
Reference: doi:10.1038/nn.4216
Source

"ctd_Zeisel2016"
CTD file with data from adult mouse hippocampus.
Reference: doi:10.1126/science.aaa1934
Source

"ctd_Zeisel2018"
CTD file with data from juvenile mouse central nervous system atlas.
Reference: doi:10.1016/j.cell.2018.06.021
Source

"ctd_LaManno2020"
CTD file with data from developing mouse central nervous system atlas.
Reference: doi:10.1038/s41586-021-03775-x
Source

"ctd_Aerts2018"
CTD file with data from aging fly central nervous system atlas.
Reference: doi:10.1016/j.cell.2018.05.057
Source

"ctd_FlyCellAtlas"
CTD file with data from adult fly whole-body atlas (a.k.a the FlyCellAtlas).
Reference: doi:10.1126/science.abk2432
Source

"ctd_Raj2018"
CTD file with data from juvenile zebrafish central nervous system atlas.
Reference: doi:10.1038/nbt.4103
Source

"ctd_Raj2020"
CTD file with data from developing zebrafish central nervous system atlas.
Reference: doi:10.1016/j.neuron.2020.09.023
Source

"ctd_Farnsworth2020"
CTD file with data from developing zebrafish central nervous system atlas.
Reference: doi:10.1016/j.ydbio.2019.11.008
Source

"ctd_TabulaMuris_facs"
CTD file with data from adult mouse whole-body atlas. Uses the Smart-seq2 (FACS) subset of the data and covers a greater number of cell-types than the Droplet version.
Reference: doi:10.1038/s41586-018-0590-4
Source

"ctd_TabulaMuris_droplet"
CTD file with data from adult mouse whole-body atlas. Uses the Droplet subset of the data
Reference: doi:10.1038/s41586-018-0590-4
Source

"ctd_TabulaMurisSenis" (Under construction)
CTD file with data from aging adult mouse whole-body atlas.
Reference: doi:10.1038/s41586-020-2496-1
Source

"ctd_Avalos2019"
CTD file with data from developing fly central nervous system atlas.
Reference: doi:10.7554/eLife.50354
Source

"ctd_NonHumanPrimateCellAtlas"
CTD file with data from adult monkey whole-body atlas.
Reference: doi:10.1038/s41586-022-04587-3
Source

"ctd_DescartesMouse"
CTD file with data from the Descartes developing mouse whole-body atlas.
Reference: doi:10.1038/s41586-019-0969-x
Source

get_ctd(
  ctd_name = c("ctd_AIBS", "ctd_allKI", "ctd_BlueLake2018_FrontalCortexOnly",
    "ctd_BlueLake2018_VisualCortexOnly", "ctd_DivSeq", "ctd_DRONC_human",
    "ctd_DRONC_mouse", "ctd_Saunders", "ctd_Tasic", "ctd_Zeisel2015", "ctd_Zeisel2018",
    "ctd_TabulaMuris_facs", "ctd_TabulaMuris_droplet", "ctd_Aerts2018",
    "ctd_FlyCellAtlas", "ctd_Avalos2019", "ctd_DescartesHuman", "ctd_DescartesMouse",
    "ctd_HumanCellLandscape", "ctd_Jiang2021", "ctd_NonHumanPrimateCellAtlas",
    "ctd_Raj2018", "ctd_Raj2018", "ctd_Farnsworth2020"),
  tag = "v2.0.8",
  .token = gh::gh_token(),
  storage_dir = tools::R_user_dir(package = "MAGMA.Celltyping", which = "cache")
)

Arguments

ctd_name

Name of one or more CellTypeDatset objects to import.

tag

tag for the GitHub release to which this data should be attached.

.token

GitHub authentication token, see [gh::gh_token()]

storage_dir

Folder in which to save the CellTypeDatset.

Value

CellTypeDataset object, or a named list of CellTypeDataset objects.

Examples

ctd <- MAGMA.Celltyping::get_ctd("ctd_AIBS")
#> Downloading 1 CellTypeDataset file(s).
#> ℹ All local files already up-to-date!