Map SNPs to their nearby genes

Make two external calls to MAGMA. First use it to annotate SNPs onto their neighbouring genes. Second, use it to calculate the gene level trait association.

map_snps_to_genes(
  path_formatted,
  genome_build = NULL,
  upstream_kb = 35,
  downstream_kb = 10,
  N = NULL,
  genome_ref_path = NULL,
  population = "eur",
  genes_only = FALSE,
  storage_dir = tools::R_user_dir("MAGMA.Celltyping", which = "cache"),
  force_new = FALSE,
  version = NULL,
  verbose = TRUE
)

Arguments

path_formatted

Filepath of the summary statistics file (which is expected to already be in the required format). Can be uncompressed or compressed (".gz" or ".bgz").

genome_build

The build of the reference genome ("GRCh37" or "GRCh38"). If NULL, it will be inferred with get_genome_build.

upstream_kb

How many kilobases (kb) upstream of the gene should SNPs be included?

downstream_kb

How many kilobases (kb) downstream of the gene should SNPs be included?

N

What is the N number for this GWAS? That is cases + controls.

genome_ref_path

Path to the folder containing the 1000 genomes reference (downloaded with get_genome_ref).

population

Which population subset of the genome reference to include.

"eur" : European descent (Default simply because this is currently the most common GWAS subpopulation).

"afr" : African descent.

"amr" : Ad Mixed American descent.

"eas" : East Asian descent.

"sas" : South Asian descent.

genes_only

The .genes.raw file is the intermediary file that serves as the input for subsequent gene-level analyses. To perform only a gene analysis, with no subsequent gene-set analysis, the --genes-only flag can be added (TRUE). This suppresses the creation of the .genes.raw file, and significantly reduces the running time and memory required.

storage_dir

Where to store genome ref.

force_new

Set to TRUE to rerun MAGMA even if the output files already exist. (Default: FALSE).

version

MAGMA version to use.

verbose

Print messages.

Value

Path to the genes.out file.

Examples

if (FALSE) {
path_formatted <- MAGMA.Celltyping::get_example_gwas()
genesOutPath <- MAGMA.Celltyping::map_snps_to_genes(
    path_formatted = path_formatted,
    genome_build = "hg19",
    N = 5000)
}